all examples, final touches
